+++ config/initializers/devise.rb
@@ -157,9 +157,7 @@ Devise.setup do |config|
   # :authlogic_sha512 (then you should set stretches above to 20 for default behavior)
   # and :restful_authentication_sha1 (then you should set stretches to 10, and copy
   # REST_AUTH_SITE_KEY to pepper)
-  if APP_CONFIG.use_asi_encryptor
-    config.encryptor = :asi
-  end
+  # config.encryptor = :encryptor
 
   # ==> Scopes configuration
   # Turn scoped views on. Before rendering "sessions/new", it will first check for