+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20160223084741) do
+ActiveRecord::Schema.define(version: 20160229114242) do
 
   create_table "auth_tokens", force: :cascade do |t|
     t.string   "token",            limit: 255
@@ -865,6 +865,7 @@ ActiveRecord::Schema.define(version: 20160223084741) do
     t.string   "username",                           limit: 255
     t.string   "email",                              limit: 255
     t.string   "encrypted_password",                 limit: 255,   default: "",    null: false
+    t.string   "legacy_encrypted_password",          limit: 255
     t.string   "reset_password_token",               limit: 255
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"