+++ spec/models/person_spec.rb
@@ -15,6 +15,7 @@
 #  username                           :string(255)
 #  email                              :string(255)
 #  encrypted_password                 :string(255)      default(""), not null
+#  legacy_encrypted_password          :string(255)
 #  reset_password_token               :string(255)
 #  reset_password_sent_at             :datetime
 #  remember_created_at                :datetime